+++ ActionSheet/index.js
@@ -1,12 +1,13 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
 import {
   Platform,
   View,
   Text,
   TouchableHighlight,
 } from 'react-native';
-import Sheet from '../Sheet';
+import PropTypes from 'prop-types';
 
+import Sheet from '../Sheet';
 import styles from './styles';
 
 const NOOP = () => {};
@@ -19,7 +20,9 @@ class ActionSheet extends Component {
         visible={this.props.visible}
         overlayStyle={this.props.overlayStyle}
         onPressOverlay={this.props.onClose}
-        onClose={this.props.onClose}
+        overlayAnimationDuration={this.props.overlayAnimationDuration}
+        onShow={this.props.onShow}
+        onHide={this.props.onHide}
         duration={this.props.duration}
         style={[styles.containerStyle, this.props.style]}
       >
@@ -103,10 +106,16 @@ ActionSheet.propTypes = {
   cancelBtnTextStyle: Text.propTypes.style,
   // 遮罩层样式
   overlayStyle: Sheet.propTypes.overlayStyle,
-  // 关闭回调（动画结束时）
-  onClose: Sheet.propTypes.onClose,
+  // Overlay 动画时长
+  overlayAnimationDuration: PropTypes.number,
+  // 关闭点击事件回调
+  onClose: Sheet.propTypes.onPressOverlay,
   // 动画时长
   duration: Sheet.propTypes.duration,
+  // 显示回调
+  onShow: PropTypes.func,
+  // 隐藏回调
+  onHide: PropTypes.func,
   // 自定义样式
   style: View.propTypes.style,
   // 按钮点击透明度变化