+++ Example/src/page/Drop/CartReceiver/index.js
@@ -1,6 +1,4 @@
-import React, {
-  PropTypes,
-} from 'react';
+import React from 'react';
 import {
   StyleSheet,
 } from 'react-native';
@@ -8,6 +6,8 @@ import Badge from 'rnx-ui/Badge';
 import {
   Receiver,
 } from 'rnx-ui/Drop';
+import PropTypes from 'prop-types';
+
 import Cart from '../Cart';
 
 const NOOP = () => {};