+++ HeaderedSheet/index.js
@@ -1,10 +1,11 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
 import {
   View,
   Text,
   TouchableOpacity,
   TouchableWithoutFeedback,
 } from 'react-native';
+import PropTypes from 'prop-types';
 
 import {
   ACTIVE_OPACITY,
@@ -66,7 +67,9 @@ class HeaderedSheet extends Component {
         visible={this.props.visible}
         overlayStyle={this.props.overlayStyle}
         onPressOverlay={this.props.onPressOverlay}
-        onClose={this.props.onClose}
+        overlayAnimationDuration={this.props.overlayAnimationDuration}
+        onShow={this.props.onShow}
+        onHide={this.props.onHide}
         duration={this.props.duration}
         style={[styles.containerStyle, this.props.containerStyle]}
       >
@@ -115,12 +118,16 @@ HeaderedSheet.propTypes = {
   visible: Sheet.propTypes.visible,
   // 遮罩层样式
   overlayStyle: Sheet.propTypes.overlayStyle,
-  // 关闭回调（动画结束时）
-  onClose: Sheet.propTypes.onClose,
   // 遮罩点击事件
   onPressOverlay: Sheet.propTypes.onPressOverlay,
+  // Overlay 动画时长
+  overlayAnimationDuration: PropTypes.number,
   // 动画时长
   duration: Sheet.propTypes.duration,
+  // 显示回调
+  onShow: PropTypes.func,
+  // 隐藏回调
+  onHide: PropTypes.func,
   // 自定容器义样式（包含 header 区域）
   containerStyle: Sheet.propTypes.style,
   // 自定义样式
@@ -153,7 +160,6 @@ HeaderedSheet.propTypes = {
 HeaderedSheet.defaultProps = {
   visible: false,
   overlayStyle: null,
-  onClose: NOOP,
   onPressOverlay: NOOP,
   duration: 200,
   style: null,