+++ ImgPicker/ImgDisplayer/index.js
@@ -1,6 +1,5 @@
 import React, {
   Component,
-  PropTypes,
 } from 'react';
 import {
   View,
@@ -9,6 +8,7 @@ import {
   TouchableHighlight,
   Text,
 } from 'react-native';
+import PropTypes from 'prop-types';
 
 import styles from './styles';