+++ ImgRollView/utils.js
@@ -1,5 +1,6 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
 import { View } from 'react-native';
+import PropTypes from 'prop-types';
 
 import style from './styles';