+++ NumericKeyboard/Key/index.js
@@ -1,5 +1,4 @@
 import React, {
-  PropTypes,
   Component,
 } from 'react';
 import {
@@ -7,6 +6,7 @@ import {
   TouchableHighlight,
   Text,
 } from 'react-native';
+import PropTypes from 'prop-types';
 
 import styles, {
   DISABLED_COLOR,