+++ TabBar/index.js
@@ -1,5 +1,4 @@
 import React, {
-  PropTypes,
   Component,
 } from 'react';
 import {
@@ -7,6 +6,8 @@ import {
   View,
   TouchableOpacity,
 } from 'react-native';
+import PropTypes from 'prop-types';
+
 import {
   HAIRLINE_WIDTH,
 } from '../constant';